feat: implement multi-tenancy, campaign metadata, and monetary precision
- Implemented multi-workspace (tenant) isolation across PostgreSQL and ClickHouse.
- Migrated all monetary fields (Cost, Payout, Revenue) to integer cents (int64/Int64) for absolute precision.
- Added comprehensive campaign metadata including grouping, cost models, and tags.
- Enhanced analytics ingestion with JA3 fingerprints, behavioral scores, and request tracing.
- Refactored repository layer with a unified DB interface for transactional safety.
- Updated Admin API handlers and click pipeline to support scoped entities and new data types.
- Included foundational architectural documentation (ARCH-DECISIONS.md, FEATURE-MAP.md).

+++ db/postgres/migrations/003_create_offers_landings.up.sql
@@ -12,7 +12,7 @@ CREATE TABLE offers (
   name                 VARCHAR(255)   NOT NULL,
   url                  TEXT           NOT NULL,
   affiliate_network_id UUID,
-  payout               DECIMAL(10,4)  NOT NULL DEFAULT 0,
+  payout               BIGINT         NOT NULL DEFAULT 0,
   state                VARCHAR(50)    NOT NULL DEFAULT 'active',
   created_at           TIMESTAMPTZ    NOT NULL DEFAULT NOW(),
   updated_at           TIMESTAMPTZ    NOT NULL DEFAULT NOW()
